+++ Client/Commands/Webcam.cs
@@ -39,7 +39,7 @@ namespace Client.Commands
             }
 
             // Get data length
-            var dataLength = int.Parse(ClientCommandsManager.networkManager.ReadLine());
+            var dataLength = long.Parse(ClientCommandsManager.networkManager.ReadLine());
 
             try
             {
